more user friendly log size information
git-svn-id: svn+ssh://foss.ulster.ac.uk/svn/opus/trunk@534 54dc3e3f-df17-4cd7-99f1-b1610cd3c088

+++ include/model/Log_Viewer.class.php
@@ -68,17 +68,17 @@ class Log_Viewer
     $megabyte = 1024*1024;
     $gigabyte = 1024*1024*1024;
 
-    if($filesize / $gigabyte)
+    if((int) ($filesize / $gigabyte))
     {
       $filesize_text = ($filesize / $gigabyte);
       return("$filesize_text GB");
     }
-    if($filesize / $megabyte)
+    if((int) ($filesize / $megabyte))
     {
       $filesize_text = ($filesize / $megabyte);
       return("$filesize_text MB");
     }
-    if($filesize / $kilobyte)
+    if((int) ($filesize / $kilobyte))
     {
       $filesize_text = ($filesize / $kilobyte);
       return("$filesize_text kB");
